refactor: Clippy-clean generated Rust codec emit (wire-identical)
- Route every Rust bit-range extract (flag getter, variant dispatch tag, parent-flag encode args) through a shared rust_bit_extract SSOT that drops >> 0/<< 0 no-ops, same-type as casts, and over-parens; sub-byte compose/decode and VLE widen emit minimally too
- Borrowed &[u8] encode passes self.<field> not &self.<field>; gate the flag-default literal on a non-zero accumulated value so all-zero codecs derive Default; targeted per-item allow on the deliberate DMA-alignment static-assert drift guard
- All 91 codec goldens pass cargo clippy -D warnings; forge_conformance 954/0 byte-stable; §6.2.6 committed trees swept (header-hash only)

+++ tests/forge/expected/codec_zenoh_encoding.rs
@@ -127,7 +127,7 @@ impl<'a> CodecZenohEncoding<'a> {
         }
         if let Some(_v) = self.schema_len {
         {
-            let mut _vle = _v as u64;
+            let mut _vle = _v;
             while _vle >= 0x80 {
                 w.write_u8((_vle as u8 & 0x7F) | 0x80)?;
                 _vle >>= 7;
