+++ tests/forge/expected/codec_zenoh_ext_zint.rs
@@ -74,7 +74,7 @@ impl CodecZenohExtZint {
         // `vle_encode_block` with the language-appropriate self/
         // struct prefix.
         {
-            let mut _vle = self.value as u64;
+            let mut _vle = self.value;
             while _vle >= 0x80 {
                 w.write_u8((_vle as u8 & 0x7F) | 0x80)?;
                 _vle >>= 7;
